+++ lib/cleanup.php
@@ -20,8 +20,6 @@ function roots_head_cleanup() {
   global $wp_widget_factory;
   remove_action('wp_head', array($wp_widget_factory->widgets['WP_Widget_Recent_Comments'], 'recent_comments_style'));
 
-  add_filter('use_default_gallery_style', '__return_null');
-
   if (!class_exists('WPSEO_Frontend')) {
     remove_action('wp_head', 'rel_canonical');
     add_action('wp_head', 'roots_rel_canonical');
@@ -127,54 +125,6 @@ function roots_body_class($classes) {
 add_filter('body_class', 'roots_body_class');
 
 /**
- * Root relative URLs
- *
- * WordPress likes to use absolute URLs on everything - let's clean that up.
- * Inspired by http://www.456bereastreet.com/archive/201010/how_to_make_wordpress_urls_root_relative/
- *
- * You can enable/disable this feature in config.php:
- * current_theme_supports('root-relative-urls');
- *
- * @author Scott Walkinshaw <scott.walkinshaw@gmail.com>
- */
-function roots_root_relative_url($input) {
-  preg_match('|https?://([^/]+)(/.*)|i', $input, $matches);
-
-  if (isset($matches[1]) && isset($matches[2]) && $matches[1] === $_SERVER['SERVER_NAME']) {
-    return wp_make_link_relative($input);
-  } else {
-    return $input;
-  }
-}
-
-function roots_enable_root_relative_urls() {
-  return !(is_admin() || in_array($GLOBALS['pagenow'], array('wp-login.php', 'wp-register.php'))) && current_theme_supports('root-relative-urls');
-}
-
-if (roots_enable_root_relative_urls()) {
-  $root_rel_filters = array(
-    'bloginfo_url',
-    'the_permalink',
-    'wp_list_pages',
-    'wp_list_categories',
-    'roots_wp_nav_menu_item',
-    'the_content_more_link',
-    'the_tags',
-    'get_pagenum_link',
-    'get_comment_link',
-    'month_link',
-    'day_link',
-    'year_link',
-    'tag_link',
-    'the_author_posts_link',
-    'script_loader_src',
-    'style_loader_src'
-  );
-
-  add_filters($root_rel_filters, 'roots_root_relative_url');
-}
-
-/**
  * Wrap embedded media as suggested by Readability
  *
  * @link https://gist.github.com/965956
@@ -184,16 +134,6 @@ function roots_embed_wrap($cache, $url, $attr = '', $post_ID = '') {
   return '<div class="entry-content-asset">' . $cache . '</div>';
 }
 add_filter('embed_oembed_html', 'roots_embed_wrap', 10, 4);
-
-/**
- * Add class="thumbnail" to attachment items
- */
-function roots_attachment_link_class($html) {
-  $postid = get_the_ID();
-  $html = str_replace('<a', '<a class="thumbnail"', $html);
-  return $html;
-}
-add_filter('wp_get_attachment_link', 'roots_attachment_link_class', 10, 1);
 
 /**
  * Add Bootstrap thumbnail styling to images with captions
@@ -233,105 +173,6 @@ function roots_caption($output, $attr, $content) {
   return $output;
 }
 add_filter('img_caption_shortcode', 'roots_caption', 10, 3);
-
-/**
- * Clean up gallery_shortcode()
- *
- * Re-create the [gallery] shortcode and use thumbnails styling from Bootstrap
- *
- * @link http://twitter.github.com/bootstrap/components.html#thumbnails
- */
-function roots_gallery($attr) {
-  $post = get_post();
-
-  static $instance = 0;
-  $instance++;
-
-  if (!empty($attr['ids'])) {
-    if (empty($attr['orderby'])) {
-      $attr['orderby'] = 'post__in';
-    }
-    $attr['include'] = $attr['ids'];
-  }
-
-  $output = apply_filters('post_gallery', '', $attr);
-
-  if ($output != '') {
-    return $output;
-  }
-
-  if (isset($attr['orderby'])) {
-    $attr['orderby'] = sanitize_sql_orderby($attr['orderby']);
-    if (!$attr['orderby']) {
-      unset($attr['orderby']);
-    }
-  }
-
-  extract(shortcode_atts(array(
-    'order'      => 'ASC',
-    'orderby'    => 'menu_order ID',
-    'id'         => $post->ID,
-    'itemtag'    => '',
-    'icontag'    => '',
-    'captiontag' => '',
-    'columns'    => 3,
-    'size'       => 'thumbnail',
-    'include'    => '',
-    'exclude'    => ''
-  ), $attr));
-
-  $id = intval($id);
-
-  if ($order === 'RAND') {
-    $orderby = 'none';
-  }
-
-  if (!empty($include)) {
-    $_attachments = get_posts(array('include' => $include, 'post_status' => 'inherit', 'post_type' => 'attachment', 'post_mime_type' => 'image', 'order' => $order, 'orderby' => $orderby));
-
-    $attachments = array();
-    foreach ($_attachments as $key => $val) {
-      $attachments[$val->ID] = $_attachments[$key];
-    }
-  } elseif (!empty($exclude)) {
-    $attachments = get_children(array('post_parent' => $id, 'exclude' => $exclude, 'post_status' => 'inherit', 'post_type' => 'attachment', 'post_mime_type' => 'image', 'order' => $order, 'orderby' => $orderby));
-  } else {
-    $attachments = get_children(array('post_parent' => $id, 'post_status' => 'inherit', 'post_type' => 'attachment', 'post_mime_type' => 'image', 'order' => $order, 'orderby' => $orderby));
-  }
-
-  if (empty($attachments)) {
-    return '';
-  }
-
-  if (is_feed()) {
-    $output = "\n";
-    foreach ($attachments as $att_id => $attachment) {
-      $output .= wp_get_attachment_link($att_id, $size, true) . "\n";
-    }
-    return $output;
-  }
-
-  $output = '<ul class="thumbnails gallery">';
-
-  $i = 0;
-  foreach ($attachments as $id => $attachment) {
-    $link = isset($attr['link']) && 'file' == $attr['link'] ? wp_get_attachment_link($id, $size, false, false) : wp_get_attachment_link($id, $size, true, false);
-
-    $output .= '<li>' . $link;
-    if (trim($attachment->post_excerpt)) {
-      $output .= '<div class="caption hidden">' . wptexturize($attachment->post_excerpt) . '</div>';
-    }
-    $output .= '</li>';
-  }
-
-  $output .= '</ul>';
-
-  return $output;
-}
-if (current_theme_supports('bootstrap-gallery')) {
-  remove_shortcode('gallery');
-  add_shortcode('gallery', 'roots_gallery');
-}
 
 /**
  * Remove unnecessary dashboard widgets
@@ -415,11 +256,11 @@ function roots_request_filter($query_vars) {
 add_filter('request', 'roots_request_filter');
 
 /**
- * Tell WordPress to use searchform.php from the templates/ directory
+ * Tell WordPress to use searchform.php from the templates/ directory. Requires WordPress 3.6+
  */
-function roots_get_search_form($argument) {
-  if ($argument === '') {
-    locate_template('/templates/searchform.php', true, false);
-  }
+function roots_get_search_form($form) {
+  $form = '';
+  locate_template('/templates/searchform.php', true, false);
+  return $form;
 }
 add_filter('get_search_form', 'roots_get_search_form');